Add task insertion during routine
Can either append or insert. The budgeted time currently defaults
to 2 minutes and can not be changed. Additions to the routine are
not saved in the routine file by design (they are meant for both
long-term adjustments and on the fly adaptions). They
will be recorded in the log when interacted with, but not if you
never elapse on them or change their status.