Reduce panic locations in code
Still some remaining. Tried to move them up to app.rs so that the
modules behave well, at least. Some remaining ones are unclear how
to handle and suggest to me that I need better models of those
areas. It also still panics on DST issues, which I will not be
prioritizing.

I think I'm okay with there being some remaining panics IF
I can mitigate data loss in the case of a panic that I had thought
would never happen. I'm not sure how to do that yet. It might be
more of a priority to work on usability than robustness right
now, considering the nature of the tool.